ebay/requirements.txt
attrs==16.0.0
cffi==1.7.0
click==6.6
cryptography==1.4
cssselect==0.9.2
enum34==1.1.6
Flask==0.11.1
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
marshmallow==2.9.0
parsel==1.0.2
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
PyYAML==3.11
queuelib==1.4.2
Scrapy==1.1.1
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.0.14
Twisted==16.3.0
ujson==1.35
w3lib==1.14.3
webargs==1.3.4
Werkzeug==0.11.10
wheel==0.24.0
zope.interface==4.2.0
